pyproject.toml
[tool.poetry]
name = "escrow-includes"
version = "0.1.1"
description = "general lib utils for escrow backend service"
authors = ["o4codes <oforkansi.shadrach@gmail.com>"]
packages = [{ include = "escrow_includes" }]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.7"
djangorestframework = "^3.14.0"
pydantic = "1.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
